pyproject.toml
[project]
name = "faceless-project"
version = "0.1.0"
description = "AI-powered video content generation platform"
readme = "readme.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "Cython==3.1.4",
    "decorator==4.4.2",
    "distro==1.9.0",
    "dtw-python==1.7.2",
    "filelock==3.20.0",
    "fsspec==2025.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "Jinja2==3.1.6",
    "jiter==0.11.1",
    "llvmlite==0.45.1",
    "MarkupSafe==3.0.3",
    "more-itertools==10.8.0",
    "moviepy==1.0.3",
    "mpmath==1.3.0",
    "networkx==3.5",
    "numba==0.62.1",
    "numpy==2.3.4",
    "openai==2.5.0",
    "Pillow==10.4.0",
    "proglog==0.1.12",
    "pydantic==2.12.3",
    "pydantic_core==2.41.4",
    "python-dotenv==1.1.1",
    "regex==2025.9.18",
    "requests==2.32.5",
    "scipy==1.16.2",
    "setuptools==80.9.0",
    "sniffio==1.3.1",
    "sympy==1.14.0",
    "tiktoken==0.12.0",
    "torch==2.9.0",
    "tqdm==4.67.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.5.0",
    "python-telegram-bot>=21.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "python-multipart>=0.0.20",
    "google-genai>=1.0.0",
    "elevenlabs>=1.0.0",
    "whisper-timestamped>=1.15.9",
    "youtube-transcript-api>=1.2.4",
    "feedparser>=6.0.12",
    "misaki[en]>=0.9.4",
    "av>=17.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cpu"]
index-strategy = "unsafe-best-match"

[dependency-groups]
dev = []
